+++ src/components/ContentPreview/ContentPreview.js
@@ -11,6 +11,8 @@ import throttle from 'lodash/throttle';
 import noop from 'lodash/noop';
 import Measure from 'react-measure';
 import PlainButton from 'box-react-ui/lib/components/plain-button/PlainButton';
+import IconNavigateLeft from 'box-react-ui/lib/icons/general/IconNavigateLeft';
+import IconNavigateRight from 'box-react-ui/lib/icons/general/IconNavigateRight';
 import ContentSidebar from '../ContentSidebar';
 import Header from './Header';
 import API from '../../api';
@@ -19,7 +21,6 @@ import Cache from '../../util/Cache';
 import makeResponsive from '../makeResponsive';
 import Internationalize from '../Internationalize';
 import { isValidBoxFile } from '../../util/fields';
-import TokenService from '../../util/TokenService';
 import { isInputElement, focus } from '../../util/dom';
 import {
     DEFAULT_HOSTNAME_API,
@@ -138,11 +139,7 @@ class ContentPreview extends PureComponent<Props, State> {
      * @return {void}
      */
     componentWillUnmount(): void {
-        if (this.preview) {
-            this.preview.removeAllListeners();
-            this.destroyPreview();
-        }
-        this.preview = undefined;
+        this.destroyPreview();
     }
 
     /**
@@ -158,11 +155,11 @@ class ContentPreview extends PureComponent<Props, State> {
         const hasFileIdChanged = nextProps.fileId !== fileId;
         const hasSizeChanged = nextProps.isSmall !== isSmall;
 
-        const newState = {};
-
         if (hasTokenChanged || hasFileIdChanged) {
-            newState.file = undefined;
             this.destroyPreview();
+            this.setState({
+                file: undefined
+            });
             this.fetchFile(nextProps.fileId);
         }
 
@@ -274,7 +271,9 @@ class ContentPreview extends PureComponent<Props, State> {
      */
     destroyPreview() {
         if (this.preview) {
+            this.preview.removeAllListeners();
             this.preview.destroy();
+            this.preview = undefined;
         }
     }
 
@@ -317,31 +316,10 @@ class ContentPreview extends PureComponent<Props, State> {
      * @param {Array<string|BoxItem>} files - files to prefetch
      * @return {void}
      */
-    async prefetch(files: Array<string | BoxItem>): Promise<void> {
-        const { token }: Props = this.props;
-
-        // We try to get tokens in bulk
-        const typedIds = files.map((file) => File.getTypedFileId(this.getFileId(file)));
-        const tokens = await TokenService.getTokens(typedIds, token);
-
+    prefetch(files: Array<string | BoxItem>): void {
         files.forEach((file) => {
             const fileId = this.getFileId(file);
-            const typedId = File.getTypedFileId(fileId);
-            const fileToken = tokens[typedId];
-            const isValidFile = isValidBoxFile(file, true, true);
-
-            if (isValidFile) {
-                const boxFile: BoxItem = ((file: any): BoxItem);
-                this.updatePreviewCacheAndPrefetch(boxFile, fileToken);
-            } else {
-                this.fetchFile(
-                    fileId,
-                    (boxfile: BoxItem) => {
-                        this.updatePreviewCacheAndPrefetch(boxfile, fileToken);
-                    },
-                    noop
-                );
-            }
+            this.fetchFile(fileId, noop, noop);
         });
     }
 
@@ -367,23 +345,17 @@ class ContentPreview extends PureComponent<Props, State> {
      */
     loadPreview = (): void => {
         const { Preview } = global.Box;
-        const { fileId, token, collection, ...rest }: Props = this.props;
+        const { token, collection, ...rest }: Props = this.props;
         const { file }: State = this.state;
 
-        if (!this.isPreviewLibraryLoaded() || !file || !token) {
+        if (!this.isPreviewLibraryLoaded() || !file || !token || this.preview) {
             return;
         }
 
-        if (!this.preview) {
-            this.preview = new Preview();
-            this.preview.addListener('load', this.onPreviewLoad);
-        }
-
-        if (this.preview.getCurrentViewer()) {
-            return;
-        }
-
-        this.preview.show(file, token, {
+        this.preview = new Preview();
+        this.preview.updateFileCache([file]);
+        this.preview.addListener('load', this.onPreviewLoad);
+        this.preview.show(file.id, token, {
             container: `#${this.id} .bcpr-content`,
             header: 'none',
             skipServerUpdate: true,
@@ -699,8 +671,10 @@ class ContentPreview extends PureComponent<Props, State> {
 
         const { file, showSidebar: showSidebarState }: State = this.state;
         const { collection }: Props = this.props;
-        const hasLeftNavigation = collection.length > 1 && this.getFileIndex() > 0;
-        const hasRightNavigation = collection.length > 1 && this.getFileIndex() < collection.length - 1;
+
+        const fileIndex = this.getFileIndex();
+        const hasLeftNavigation = collection.length > 1 && fileIndex > 0 && fileIndex < collection.length;
+        const hasRightNavigation = collection.length > 1 && fileIndex > -1 && fileIndex < collection.length - 1;
         const isValidFile = isValidBoxFile(file, true, true);
 
         let isSidebarVisible = isValidFile && hasSidebar && showSidebarState;
@@ -744,28 +718,12 @@ class ContentPreview extends PureComponent<Props, State> {
                             </Measure>
                             {hasLeftNavigation && (
                                 <PlainButton type='button' className='bcpr-navigate-left' onClick={this.navigateLeft}>
-                                    <svg viewBox='0 0 48 48' focusable='false'>
-                                        <path
-                                            fill='#494949'
-                                            stroke='#DCDCDC'
-                                            strokeMiterlimit='10'
-                                            d='M30.8,33.2L21.7,24l9.2-9.2L28,12L16,24l12,12L30.8,33.2z'
-                                        />
-                                        <path display='none' fill='none' d='M0,0h48v48H0V0z' />
-                                    </svg>
+                                    <IconNavigateLeft />
                                 </PlainButton>
                             )}
                             {hasRightNavigation && (
                                 <PlainButton type='button' className='bcpr-navigate-right' onClick={this.navigateRight}>
-                                    <svg viewBox='0 0 48 48' focusable='false'>
-                                        <path
-                                            fill='#494949'
-                                            stroke='#DCDCDC'
-                                            strokeMiterlimit='10'
-                                            d='M17.2,14.8l9.2,9.2l-9.2,9.2L20,36l12-12L20,12L17.2,14.8z'
-                                        />
-                                        <path display='none' fill='none' d='M48,48H0L0,0l48,0V48z' />
-                                    </svg>
+                                    <IconNavigateRight />
                                 </PlainButton>
                             )}
                         </div>