+++ CHANGELOG.md
@@ -1,6 +1,43 @@
+<a name="4.3.2"></a>
+## <small>4.3.2 (2018-03-15)</small>
+
+* Fix: Show skills even if one card is valid (#225) ([eacea35](https://github.com/box/box-ui-elements/commit/eacea35)), closes [#225](https://github.com/box/box-ui-elements/issues/225)
+
+
+
+<a name="4.3.1"></a>
+## <small>4.3.1 (2018-03-15)</small>
+
+* Release: 4.3.1 ([0b0c0c9](https://github.com/box/box-ui-elements/commit/0b0c0c9))
+* Fix: Flow error with latest webpack-cli (#224) ([a430cad](https://github.com/box/box-ui-elements/commit/a430cad)), closes [#224](https://github.com/box/box-ui-elements/issues/224)
+
+
+
+<a name="4.3.0"></a>
+## 4.3.0 (2018-03-14)
+
+* Release: 4.3.0 ([1456442](https://github.com/box/box-ui-elements/commit/1456442))
+* Mojito: Update translations (#219) ([0211f28](https://github.com/box/box-ui-elements/commit/0211f28)), closes [#219](https://github.com/box/box-ui-elements/issues/219)
+* Mojito: Update translations (#223) ([c26168c](https://github.com/box/box-ui-elements/commit/c26168c)), closes [#223](https://github.com/box/box-ui-elements/issues/223)
+* Chore: Minor folder structure cleanup (#222) ([74bf4c2](https://github.com/box/box-ui-elements/commit/74bf4c2)), closes [#222](https://github.com/box/box-ui-elements/issues/222)
+
+
+
+<a name="4.2.1"></a>
+## <small>4.2.1 (2018-03-14)</small>
+
+* Release: 4.2.1 ([571f7b2](https://github.com/box/box-ui-elements/commit/571f7b2))
+* Fix: Build issues (#216) ([16cd7da](https://github.com/box/box-ui-elements/commit/16cd7da)), closes [#216](https://github.com/box/box-ui-elements/issues/216)
+* Fix: Navigation issues of preview (#218) ([0503174](https://github.com/box/box-ui-elements/commit/0503174)), closes [#218](https://github.com/box/box-ui-elements/issues/218)
+* Update: Add left/right navigation for preview (#217) ([1a0ff4c](https://github.com/box/box-ui-elements/commit/1a0ff4c)), closes [#217](https://github.com/box/box-ui-elements/issues/217)
+* Mojito: Update translations (#215) ([a628fb3](https://github.com/box/box-ui-elements/commit/a628fb3)), closes [#215](https://github.com/box/box-ui-elements/issues/215)
+
+
+
 <a name="4.2.0"></a>
 ## 4.2.0 (2018-03-10)
 
+* Release: 4.2.0 ([14e777b](https://github.com/box/box-ui-elements/commit/14e777b))
 * Fix: Preview navigation hook (#214) ([00096b7](https://github.com/box/box-ui-elements/commit/00096b7)), closes [#214](https://github.com/box/box-ui-elements/issues/214)
 * New: Add download and print functionality in Preview in Explorer (#213) ([c93c045](https://github.com/box/box-ui-elements/commit/c93c045)), closes [#213](https://github.com/box/box-ui-elements/issues/213)